requirements.txt
# Human readable time deltas.
ago

# DKIM Signed Email from Python.
dkimpy
